refactor(invoices): split InvoicesPage into feature modules
